chore: Move stories files to stories folder
