+++ 김도연/week4_url-shortening/src/stories/RoundFillButton.stories.tsx
@@ -1,5 +1,5 @@
 import { Meta, StoryObj } from '@storybook/react';
-import RoundFillButton from './RoundFillButton';
+import RoundFillButton from '../components/atoms/RoundFillButton';
 
 const meta: Meta<typeof RoundFillButton> = {
   title: 'RoundFillButton',
